+++ packages/pitchset/index.js
@@ -7,13 +7,36 @@
  *
  * @module pitchset
  */
-import { chroma } from 'tonal-note'
-import { map, asArr } from 'tonal-array'
+import { chr, asPitch } from 'tonal-pitch'
+import { map, asArr, rotate, compact } from 'tonal-array'
+import { transpose } from 'tonal-transpose'
 
 function toInt (set) { return parseInt(toBinary(set), 2) }
+function chroma (p) { p = asPitch(p); return p ? chr(p) : null }
 
 /**
- * Convert a pitch set into a binary representation
+ * Get the pitchset binary rotations of a list of notes
+ */
+export function rotations (set, normalize) {
+  normalize = normalize !== false
+  var binary = toBinary(set).split('')
+  return compact(binary.map(function (_, i) {
+    var r = rotate(i, binary)
+    return normalize && r[0] === '0' ? null : r.join('')
+  }))
+}
+
+var REGEX = /^[01]{12}$/
+/**
+ * Test if the given value is a pitch set in binary representation
+ */
+export function isBinary (set) {
+  return REGEX.test(set)
+}
+
+/**
+ * Convert a pitch set into a binary representation. If the argument is
+ * already a binary representation it returns it.
  *
  * @param {Array|String} set - the pitch set
  * @return {String} a binary representation of the pitch set
@@ -21,11 +44,45 @@ function toInt (set) { return parseInt(toBinary(set), 2) }
  * pitchset.toBinary('C D E') // => '1010100000000'
  */
 export function toBinary (set) {
+  if (isBinary(set)) return set
   var b = [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
   map(chroma, set).forEach(function (i) {
     b[i] = 1
   })
   return b.join('')
+}
+
+/**
+ * Get a pitch set with different tonic
+ * @param {String|Pitch} tonic - the desired tonic
+ * @param {Array|String} set - the list of notes or the binary representation
+ * @param {Array} a list of notes or intervals (depending the root)
+ * @example
+ * pitchset.withTonic('c d e f g a b', 'D')
+ */
+export function withTonic (tonic, set) {
+  if (arguments.length === 1) return function (s) { return withTonic(tonic, s) }
+  return fromBinary(toBinary(set), tonic)
+}
+
+var IVLS = '1P 2m 2M 3m 3M 4P 5d 5P 6m 6M 7m 7M'.split(' ')
+/**
+ * Given a pitch set in binary notation it returns the intervals or notes
+ * (depending on the tonic)
+ * @param {String} binary - the pitch set in binary representation
+ * @param {String|Pitch} tonic - the pitch set tonic
+ * @return {Array} a list of notes or intervals
+ * @example
+ * pitchset.fromBinary('101010101010', 'C') // => ['C', 'D', 'E', 'Gb', 'Ab', 'Bb']
+ */
+export function fromBinary (binary, tonic) {
+  if (arguments.length === 1) return function (t) { return fromBinary(binary, t) }
+  if (!isBinary(binary)) return null
+
+  tonic = tonic || 'P1'
+  return compact(binary.split('').map(function (d, i) {
+    return d === '1' ? transpose(IVLS[i], tonic) : null
+  }))
 }
 
 /**